requirements.txt
astroid==2.8.0
attrs==21.2.0
click==8.0.1
Flask==2.0.1
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
toml==0.10.2
typing-extensions==3.10.0.2
Werkzeug==2.0.1
wrapt==1.12.1
